Made thunderdome not report attack logs
breath deeply, admins

+++ code/modules/mob/mob_grab.dm
@@ -141,9 +141,10 @@
 			icon_state = "grabbed+1"
 			if(!affecting.buckled)
 				affecting.loc = assailant.loc
-			affecting.attack_log += "\[[time_stamp()]\] <font color='orange'>Has had their neck grabbed by [assailant.name] ([assailant.ckey])</font>"
-			assailant.attack_log += "\[[time_stamp()]\] <font color='red'>Grabbed the neck of [affecting.name] ([affecting.ckey])</font>"
-			msg_admin_attack("[key_name(assailant)] grabbed the neck of [key_name(affecting)] (<A HREF='?_src_=holder;adminplayerobservejump=\ref[assailant]'>JMP</A>)")
+			if(!in_unlogged(assailant))
+				affecting.attack_log += "\[[time_stamp()]\] <font color='orange'>Has had their neck grabbed by [assailant.name] ([assailant.ckey])</font>"
+				assailant.attack_log += "\[[time_stamp()]\] <font color='red'>Grabbed the neck of [affecting.name] ([affecting.ckey])</font>"
+				msg_admin_attack("[key_name(assailant)] grabbed the neck of [key_name(affecting)] (<A HREF='?_src_=holder;adminplayerobservejump=\ref[assailant]'>JMP</A>)")
 			hud.icon_state = "disarm/kill"
 			hud.name = "disarm/kill"
 		else
@@ -162,9 +163,10 @@
 						return
 					state = GRAB_KILL
 					assailant.visible_message("<span class='danger'>[assailant] has tightened \his grip on [affecting]'s neck!</span>")
-					affecting.attack_log += "\[[time_stamp()]\] <font color='orange'>Has been strangled (kill intent) by [assailant.name] ([assailant.ckey])</font>"
-					assailant.attack_log += "\[[time_stamp()]\] <font color='red'>Strangled (kill intent) [affecting.name] ([affecting.ckey])</font>"
-					msg_admin_attack("[key_name(assailant)] strangled (kill intent) [key_name(affecting)] (<A HREF='?_src_=holder;adminplayerobservejump=\ref[assailant]'>JMP</A>)")
+					if(!in_unlogged(assailant))
+						affecting.attack_log += "\[[time_stamp()]\] <font color='orange'>Has been strangled (kill intent) by [assailant.name] ([assailant.ckey])</font>"
+						assailant.attack_log += "\[[time_stamp()]\] <font color='red'>Strangled (kill intent) [affecting.name] ([affecting.ckey])</font>"
+						msg_admin_attack("[key_name(assailant)] strangled (kill intent) [key_name(affecting)] (<A HREF='?_src_=holder;adminplayerobservejump=\ref[assailant]'>JMP</A>)")
 
 					assailant.next_move = world.time + 10
 					affecting.losebreath += 1
